Bindings as Array
This change updates Bindings to be an array of elements with a name field,
rather than a map of elements.  This makes passing around a binding, without
losing its name, easier.  This change also adds convenience methods for
getting the Kind, Provider, and Tags for a binding rather than everyone
looking them up by constant.

Signed-off-by: Ben Hale <bhale@pivotal.io>