BuildResult Containing BuildpackPlan Pointer
Previously the BuildResult contained a BuildpackPlan struct, not a pointer.
This had been fine up until this point as all population of that plan happened
before it was returned by the buildpack.  However, when implementing a bit of
functionality that added to the plan as the layer was being contributed, it
became apparent that the new data was not being written.  This change changes
the BuildResult to have a pointer to the BuildpackPlan so that it can continue
to be updated after it has been returned by the buildpack, but before it has
been read by the Build function.

Signed-off-by: Ben Hale <bhale@pivotal.io>