Single Entrypoint
Previously, the library was intended to be used by two different Go binaries,
each calling a different entry function (Build() and Detect()).  It was
pointed out that this ended up bloating the buildpacks with ~5.8M of Go
runtime for each binary.  This change adds a convenience Main() function that
allows a single binary to be built (typically symlinked from the appropriate
names) which then dispatches to the proper entry function.

Signed-off-by: Ben Hale <bhale@pivotal.io>