Supports bindings created from k8s config maps & secrets
When binding metadata or secrets are created using k8s config maps or k8s
secrets, respectively, the key files are symlinks to data files in a hidden
directory

a simplified example:
├── metadata
│   ├── .hidden
│   │   └── test-metadata-key
│   └── test-metadata-key -> .hidden/test-metadata-key
└── secret
    ├── .hidden
    │   └── test-secret-key
    └── test-secret-key -> .hidden/test-secret-key

Therefore, when parsings binding we should not fail on the existance of
directories in the metadata & secret dirs.

In addition we will ignore all hidden files as a general defensive
posture.

* also go fmt

Signed-off-by: Emily Casey <ecasey@pivotal.io>