Update Environment Var APIs
Given an upcoming spec change around how environment variables are written to
the filesystem, this change updates the API to be compatible with both old and
new versions.  It does this by eliminating the "unqualified" API and requiring
explicit declarations for all API usage.  It also eliminates the "delimiter"
APIs in favor of requiring a delimiter for all "append" and "prepend" APIs.

[buildpacks/spec#131][buildpacks/rfcs#98]

Signed-off-by: Ben Hale <bhale@vmware.com>