Removes the Passed field from DetectResult
- This boolean is an unneeded structure that can be accomplished with
the addition of libcnb.Fail which is a way for a buildpack to create an
error that the error handler will interpret as a detection failure as
opposed to a detection error. This function also allows for the addition
of messages that can clue the user in on why detection failed.

Signed-off-by: Forest Eckhardt <feckhardt@pivotal.io>