Finish removing shell-specific logic (RFC #0093)
This PR submits the second half of RFC #0093, which is removing the Direct field from the Process struct and also switching the Command field from a string to a list of strings.

This is a breaking change as it removes a field from the struct and changes an existing field, but given this change is only going into v2+ that seems to be the correct path.

Resolves #70.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>