Extract method for generating Config (#151)
* Extract method for generating Config

This PR extracts the code to create a Config object from a list of applied Options. The method also allows us to enforce a global set of default options.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>

* Switch to passinig Logger through the context

It was mentioned that changing the DetectFunc and BuildFunc creates an undesirable amount of breakage in existing buildpack. Instead of passing the logger that way, this PR now passes the logger through the DetectContext and BuildContext which is already passed through to DetectFunc and BuildFunc.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>

* NewConfigWithOptions -> NewConfig

Signed-off-by: Sambhav Kothari <skothari44@bloomberg.net>

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>
Signed-off-by: Sambhav Kothari <skothari44@bloomberg.net>
Co-authored-by: Daniel Mikusa <dmikusa@vmware.com>
Co-authored-by: Sambhav Kothari <skothari44@bloomberg.net>