Make buildpack API generic
Instead of using `interface{}` for things like metadata, we can use generic types. The buildpack API doesn't care what's in those values, so it can just be an `any` type. This will allow buildpack authors to set the types they want to use and even use custom types for their metadata. Then when accessing the metadata in their buildpacks, no casting is required. This reduces the amount of code and makes the code less error prone.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>